feat: Eliminate dependency on gcc
This commit adds the following change to remove the dependency on gcc:

* Rewrite `syscall` as a naked function
* Introduce a new feature `dlopen-c` which controls the presence of C
based dlopen. This feature is enabled by default so as not to
introduce any breaking changes

The result here, is that we can have a pure Rust ckb-std crate without
dependency on C compilers, which can be suitable for more use cases.